move duplicate_index_ordering to read_heterogeneous_graph
